feat: adjust folders to paths
BREAKING CHANGE: context behavir changed

The handler context now contains `case|baseline|resultPath` instead of folder.

The behavior for file-case and folder-case now are are same.

`casePath` points to the actual case file or folder.
`baseline|resultPath` always points to a folder  with case name.
Even if it is a file-case,
a folder with that name (including file extension) will be created.

This allows user to simply check `caseType` to see if it is a file or folder case and read the right thing,
based on their assumption of the structure.

This allows the `baseline()` to be used in a fixture folder with mixed file and folder cases.